fix: check ReparseRange before reading it (#5)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰
v    Before smashing the submit button please review the checkboxes.
v If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Description
This PR fixes a bug that cause the crash of the parser when the
`ReparseRange` field is not defined.

## Checklist
- [ ] Targeted PR against correct branch.
- [ ] Linked to Github issue with discussion and accepted design OR link
to spec that describes this work.
- [ ] Wrote unit tests.
- [ ] Re-reviewed `Files changed` in the Github PR explorer.

(cherry picked from commit bc1b4e543fda0f5281f8fdea050160bb770efd7a)