refactor!: add context to the root command   (#9)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Description

This PR adds a new `Context` to the root command generated by the
`cmdtypes.RootCmd` function. This new `Context` will provide the parsing
configuration, the database instance, the node from which we parse the
data, and all the modules registered by the users. With all this data, a
client using the Juno library will be able to access the various modules
to create custom commands that utilize one or more of these modules.

## Checklist
- [x] Targeted PR against correct branch.
- [ ] Linked to Github issue with discussion and accepted design OR link
to spec that describes this work.
- [ ] Wrote unit tests.  
- [x] Re-reviewed `Files changed` in the Github PR explorer.

---------

Co-authored-by: Billie <151112678+rustcandy@users.noreply.github.com>